requirements.txt
dash
emoji
google-auth
ipdb
pandas
pandas-gbq
plotly-express
python-dotenv
requests
survey
